Deleted app folder, didn't need
-Also added put endpoints
